backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
psycopg2-binary==2.9.9
pylint==3.0.2
python-dotenv==1.0.0
Pillow==10.1.0
djoser==2.2.0
djangorestframework-simplejwt==5.3.0
textdistance==4.5.0
difflib3==0.1.2
